Versão 28.05.16.1

+++ app/src/main/java/com/pap/queropizza3/models/TPedidoDetalhe.java
@@ -8,7 +8,7 @@ public class TPedidoDetalhe {
     int id_detalhe;
     int id_item;
     TCardapioItem cardapio_item;
-    float valor;
+    double valor;
 
     public int getId_detalhe() {
         return id_detalhe;
@@ -34,11 +34,11 @@ public class TPedidoDetalhe {
         this.cardapio_item = cardapio_item;
     }
 
-    public float getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(float valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
 }
